pm_os/requirements.txt
anthropic>=0.39.0
gradio>=4.0.0
chromadb>=0.4.0
networkx>=3.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
